+++ common/src/main/java/io/github/mortuusars/exposure/item/FilmRollItem.java
@@ -4,12 +4,20 @@ import io.github.mortuusars.exposure.Config;
 import io.github.mortuusars.exposure.PlatformHelper;
 import io.github.mortuusars.exposure.camera.infrastructure.FilmType;
 import io.github.mortuusars.exposure.gui.ClientGUI;
+import io.github.mortuusars.exposure.menu.ItemRenameMenu;
 import net.minecraft.ChatFormatting;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
 import net.minecraft.nbt.Tag;
 import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.MenuProvider;
+import net.minecraft.world.entity.player.Inventory;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
@@ -85,6 +93,9 @@ public class FilmRollItem extends Item implements IFilmItem {
                             .withStyle(ChatFormatting.GRAY));
         }
 
+        if (Config.Common.FILM_ROLL_RENAMING.get()) {
+            tooltipComponents.add(Component.translatable("item.exposure.film_roll.tooltip.renaming"));
+        }
 
         // Create compat:
         int developingStep = stack.getTag() != null ? stack.getTag().getInt("CurrentDevelopingStep") : 0;
@@ -105,8 +116,40 @@ public class FilmRollItem extends Item implements IFilmItem {
                     .withStyle(ChatFormatting.GOLD));
         }
 
+        //noinspection ConstantValue
         if (exposedFrames > 0 && !PlatformHelper.isModLoaded("jei") && Config.Client.RECIPE_TOOLTIPS_WITHOUT_JEI.get()) {
             ClientGUI.addFilmRollDevelopingTooltip(stack, level, tooltipComponents, isAdvanced);
         }
+    }
+
+    @Override
+    public @NotNull InteractionResultHolder<ItemStack> use(Level level, Player player, InteractionHand usedHand) {
+        if (!Config.Common.FILM_ROLL_RENAMING.get() || !(player instanceof ServerPlayer serverPlayer)) {
+            return super.use(level, player, usedHand);
+        }
+
+        int slot = getMatchingSlotInInventory(player.getInventory(), player.getItemInHand(usedHand));
+        MenuProvider menuProvider = new MenuProvider() {
+            @Override
+            public @NotNull Component getDisplayName() {
+                return Component.translatable("gui.exposure.item_rename.title");
+            }
+
+            @Override
+            public @NotNull AbstractContainerMenu createMenu(int containerId, @NotNull Inventory playerInventory, @NotNull Player player) {
+                return new ItemRenameMenu(containerId, playerInventory, slot);
+            }
+        };
+        PlatformHelper.openMenu(serverPlayer, menuProvider, buffer -> buffer.writeInt(slot));
+        return InteractionResultHolder.success(player.getItemInHand(usedHand));
+    }
+
+    protected int getMatchingSlotInInventory(Inventory inventory, ItemStack stack) {
+        for (int i = 0; i < inventory.getContainerSize(); i++) {
+            if (inventory.getItem(i).equals(stack)) {
+                return i;
+            }
+        }
+        return -1;
     }
 }